backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
python-dotenv
pydantic-settings
pgvector
boto3>=1.34.0
mistralai>=0.1.0
pytest
pytest-asyncio
httpx
pypdf
python-multipart
pytest-mock
langgraph
langchain
openai
tiktoken
python-multipart
